backend/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.8.0